src/access/requirements.txt
nameko==2.12.0
pyjwt==1.7.1